Part 2 solution

+++ day08/src/lib.rs
@@ -24,7 +24,7 @@ pub mod initial;
 pub use crate::initial::Day08Initial;
 
 type Day08SolutionPart1 = usize;
-type Day08SolutionPart2 = i64;
+type Day08SolutionPart2 = usize;
 
 pub trait AoC<'a>: Debug {
     type SolutionPart1;
